move encoder to highlighter namespace
