+++ src/Highlighter/Replacer/EncodedReplacer.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace VStelmakh\UrlHighlight\Highlighter\Replacer;
 
+use VStelmakh\UrlHighlight\Highlighter\Encoder\EntityDecoder;
 use VStelmakh\UrlHighlight\Highlighter\Linker\Linker;
 use VStelmakh\UrlHighlight\Highlighter\Tokenizer\Token\TagToken;
 use VStelmakh\UrlHighlight\Highlighter\Tokenizer\Tokenizer;
-use VStelmakh\UrlHighlight\Matcher\EntityDecoder;
 use VStelmakh\UrlHighlight\Matcher\Matcher;
 use VStelmakh\UrlHighlight\Matcher\UrlMatch;
